kernel-module-gles: disable enum-conversion error for GCC 13 compatibility
While migrating from Yocto Kirkstone to Scarthgap, the GSX kernel
module failed to build with GCC 13 due to stricter
enum type checking:

    error: implicit conversion from 'enum <anonymous>'
    to 'IMG_BOOL' {aka 'enum tag_img_bool'} [-Werror=enum-conversion]
    |  1349 |                         bEnableAPM = false;
    |       |

This issue occurs because the prebuilt graphics sources use
'true/false' instead of 'IMG_TRUE/IMG_FALSE' for IMG_BOOL values.
Since -Werror is enabled in the module’s Makefile, the warning
is treated as a build error.

As a temporary workaround, this change appends the following flags to
disable the error:

    EXTRA_OEMAKE:append = ' KBUILD_CFLAGS+=" -Wno-error=enum-conversion" \
                            KCFLAGS+=" -Wno-error=enum-conversion"'

This allows the build to proceed on Scarthgap until the graphics package
is updated or patched accordingly.

Signed-off-by: Mykhailo Androsiuk <mykhailo_androsiuk@epam.com>