Stopping support for `salvator-xs-m3-2x4g`
While transitioning from Yocto Kirkstone to Scarthgap, it was decided
to review which boards we still support.

– Based on this source:
https://elinux.org/R-Car/Boards/Yocto-Gen3/v5.9.7
it was decided to drop the very specific `salvator-xs-m3-2x4g` board.
It’s outdated and was never used in our project.

– The following boards are under review:
	- `salvator-xs-h3-4x2g`
	- `salvator-x-h3-4x2g`
Maybe we should stop supporting them as well. This is still being
discussed.

Also, in `prod-devel-rcar.yaml`, the `pack-ipl-m3` section was removed
because it is no longer relevant.

Signed-off-by: Mykhailo Androsiuk <mykhailo_androsiuk@epam.com>