prod-devel-rcar-virtio.yaml: align domu config with new Poky naming scheme
During the migration from Kirkstone to Scarthgap, we discovered
a behavioral change in Poky:
newer releases append .rootfs not only to the image filename,
but also to reference the artifact.
For domd, this change is leveled out inside the
"../meta-renesas/meta-rcar-gen3" layer, so no updates are required.
For domu, however, the product YAML still referenced the pre-scarthgap
image path, which caused full.img to depend on a non-existing artifact
and resulted in a Ninja error.

This commit updates the domu configuration to comply with the new Poky
naming scheme.

Signed-off-by: Mykhailo Androsiuk <mykhailo_androsiuk@epam.com>